dev_requirements.txt
appdirs==1.4.4
asgiref==3.4.1
astroid==2.6.2
black==21.6b0
click==8.0.1
Django==3.2.5
isort==5.9.1
lazy-object-proxy==1.6.0
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.8.1
Pillow==8.3.0
pkg-resources==0.0.0
pylint==2.9.3
pytz==2021.1
regex==2021.7.6
sqlparse==0.4.1
toml==0.10.2
Wand==0.6.6
wrapt==1.12.1
linkify-it-py==1.0
markdown-it-py==1.1.0
markdown-it-py[linkify==1.1.0
